+++ test/test_kojo/subscribers/internal.subsA.js
@@ -4,7 +4,7 @@ const sleep = require('util').promisify(setTimeout);
 
 module.exports = async (kojo, logger) => {
 
-    const {alpha, bravo} = kojo.modules;
+    const {alpha, bravo} = kojo.services;
     assert(alpha.methodA);
     assert(bravo.methodA);
     assert(typeof logger.debug === 'function');